feat: ♻️ Refactor binding strategy for City-based views to improve modularity and testability
This commit refactors the way City bindings are handled across multiple views:

- ✅ Converted `@Binding var city` to `let city: Binding<City>` in:
  • CityInfoCard
  • FavoriteToggleButton
  • SearchRowView

- 🚫 Kept `@Binding var city` in `CityDetailView`, since it orchestrates navigation and manages the binding directly.

Benefits:
- Aligns with SOLID principles (SRP, DIP)
- Improves reusability and flexibility of subcomponents
- Enhances testability and preview support
- Reduces tight coupling to SwiftUI binding lifecycle

All existing features remain functional and UI identifiers are preserved for testing.
