+++ Smart City/Features/CityDetail/ViewModels/CityDetailViewModel.swift
